backend/requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.6
sqlalchemy>=2.0.35
pydantic>=2.9.2
python-multipart>=0.0.12
httpx>=0.27.2
chromadb>=0.5.23
PyPDF2>=3.0.1
python-docx>=1.1.2
